requirements.txt
anthropic
claude-agent-sdk
anyio
pyyaml
